requirements.txt
openai>=1.30.0
pymilvus>=2.4.0
sentence-transformers>=3.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
hdbscan>=0.8.36
pandas>=2.2.0
python-dotenv>=1.0.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
tqdm>=4.66.0
